requirements.txt
altgraph==0.17.4
esper==2.5
packaging==24.2
pefile==2023.2.7
pygame-ce==2.5.3
pygbag==0.9.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pywin32-ctypes==0.2.3
